move sitter auth  from auth slice to new slice
